Add Ability to Overwrite Symbolic Links With the --overwritesymlinks option enabled, mmirror will remove and link existing destinations. This only applies if the existing destinations are links not regular folders.

+++ mmirror.py
@@ -50,10 +50,12 @@ class Folder(object):
               'symlinks will be created. 1 will link folders under source.')
 @click.option('--followsymlinks', is_flag=True,
               help='Follow symbolic links in the source paths')
+@click.option('--overwritesymlinks', is_flag=True,
+              help='Overwrite symlinks in the output directory.')
 @click.option('-v', '--verbose', count=True,
               help='Logging verbosity, -vv for very verbose.')
 def mmirror(source_high, source_low, output_high, output_low, depth,
-            followsymlinks, verbose):
+            followsymlinks, overwritesymlinks, verbose):
     """Create a symlinked merged directory of high and low sources.
 
     Two inputs are provided which have some overlapping data but with different
@@ -70,6 +72,10 @@ def mmirror(source_high, source_low, output_high, output_low, depth,
     correspond to which source is used.
     Low: l0, l1, l2, l3, l4, h5
     High: l0, l1, h2, h3, h4, 5
+
+    To update the output directories, run with --overwritesymlinks. This will
+    update the symbolic links with new values. Be careful, this should be run
+    with the same depth value as the first run.
     """
     if verbose == 1:
         log.setLevel(logging.INFO)
@@ -102,10 +108,10 @@ def mmirror(source_high, source_low, output_high, output_low, depth,
     log.debug('Dumping low source list\n%s', pformat(slow))
 
     if output_high:
-        mirror(output_high, shigh, slow)
+        mirror(output_high, shigh, slow, overwritesymlinks)
 
     if output_low:
-        mirror(output_low, slow, shigh)
+        mirror(output_low, slow, shigh, overwritesymlinks)
 
 
 def iterate_input(absolute, depth, followsymlinks, relative=''):
@@ -131,7 +137,7 @@ def iterate_input(absolute, depth, followsymlinks, relative=''):
     return result
 
 
-def mirror(path, primary, secondary):
+def mirror(path, primary, secondary, overwrite):
     """Merge secondary onto primary and output the result to path.
 
     First, ensure the target path exists. Then merge the two lists of Folders
@@ -142,20 +148,15 @@ def mirror(path, primary, secondary):
         log.info('Creating directory %s', path)
         os.mkdir(path)
 
-    # Overwrite options are not clear. Fail if the output directory isn't empty
-    output = iterate_input(path, 1, False)
-    if len(output):
-        raise Exception('Output directory must be empty.')
-
     merged = list(primary)
     merged.extend(filter(lambda x: x not in primary, secondary))
     merged.sort(key=Folder.relative_path)
     log.debug('Dumping merged list\n%s', pformat(merged))
 
-    create_output(path, merged)
+    create_output(path, merged, overwrite)
 
 
-def create_output(base, folders):
+def create_output(base, folders, overwrite):
     """Create the output directory structure.
 
     At the correct level of folder items create the symlink structure. The
@@ -164,8 +165,21 @@ def create_output(base, folders):
     for f in folders:
         path = '%s/%s' % (base, f.relative_path)
         if f.at_depth:
-            log.debug('Linking %s to %s', path, f.absolute_path)
-            os.symlink(f.absolute_path, path)
+            # This is the item that needs to be linked. If the destination does
+            # not exist, link it. If it does, one of two things can happen: if
+            # the destination is a link and overwritesymlinks is set, remove it
+            # then link it; otherwise, do nothing.
+            if os.path.exists(path):
+                if os.path.islink(path) and overwrite:
+                    os.remove(path)
+                    log.debug('Linking (overwriting) %s to %s', path,
+                              f.absolute_path)
+                    os.symlink(f.absolute_path, path)
+                else:
+                    log.info('Not overwriting %s', path)
+            else:
+                log.debug('Linking %s to %s', path, f.absolute_path)
+                os.symlink(f.absolute_path, path)
         elif not os.path.exists(path):
             log.debug('Creating directory: %s', path)
             os.mkdir(path)
